+++ src/bl4ckscor3/bot/bl4ckb0t/misc/ShowTweet.java
@@ -21,7 +21,7 @@ public class ShowTweet
 	 * @param link The link to the Tweet
 	 * @param depth Recursive depth
 	 */
-	public static void show(String channel, String link, int depth) throws IOException
+	public static void show(String channel, String link, String user, int depth) throws IOException
 	{
 		String name = "";
 		String account = "";
@@ -72,7 +72,7 @@ public class ShowTweet
 		}
 		
 		account = "@" + doc.select("a.account-group:nth-child(2) > span:nth-child(4) > b:nth-child(2)").get(0).toString().replace("<b>", "").replace("</b>", "");
-		tweet = doc.select(".TweetTextSize--26px").get(0).text().replace("https://twitter.com", " https://twitter.com").replace("pic.twitter", " pic.twitter").trim();
+		tweet = doc.select(".TweetTextSize--26px").get(0).text().replace("https://twitter.com", " https://twitter.com").replace("pic.twitter", " pic.twitter").replace(" ", "").trim();
 		
 		String msg = Colors.BOLD + name + " (" + account + ") - " + Colors.BOLD + tweet;
 		
@@ -99,10 +99,11 @@ public class ShowTweet
 		{
 			for(String s : tweet.split(" "))
 			{
-				if(s.endsWith(" …"))
-					show(channel, "https://twitter.com" + s.split("https://twitter.com")[1], ++depth);
+				if(s.contains("twitter.com"))
+					show(channel, s, user, depth + 1);
 			}
 		}
 		catch(ArrayIndexOutOfBoundsException e){} //happens when the Tweet doesn't actually have any quoted Tweet in it
+		LinkManager.handleLink(tweet, channel, user, true);
 	}
 }